requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-spanner==3.35.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
jsondiff==2.0.0
logger==1.4
mysql-connector-python==8.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
Jinja2==3.1.2
